Added API & call etc formweb2mail pref
