+++ src/context/talentLayer.tsx
@@ -7,8 +7,8 @@ import { getUserByAddress } from '../queries/users';
 import { iTalentLayerContext, IUser } from '../types';
 import { getCompletionScores, ICompletionScores } from '../utils/profile';
 import { getWorkerProfileByOwnerId } from '../modules/BuilderPlace/request';
-import { IWorkerProfile } from '../modules/BuilderPlace/types';
 import { MAX_TRANSACTION_AMOUNT } from '../config';
+import { User } from '.prisma/client';
 
 const TalentLayerContext = createContext<iTalentLayerContext>({
   loading: true,
@@ -25,10 +25,11 @@ const TalentLayerProvider = ({ children }: { children: ReactNode }) => {
   const chainId = useChainId();
   const { data: walletClient } = useWalletClient();
   const [user, setUser] = useState<IUser | undefined>();
-  const [workerProfile, setWorkerProfile] = useState<IWorkerProfile | undefined>();
+  const [workerProfile, setWorkerProfile] = useState<User>();
   const account = useAccount();
   const [canUseDelegation, setCanUseDelegation] = useState(false);
   const [loading, setLoading] = useState(true);
+  const [workerDataLoading, setWorkerDataLoading] = useState(false);
   const [completionScores, setCompletionScores] = useState<ICompletionScores | undefined>();
   const [talentLayerClient, setTalentLayerClient] = useState<TalentLayerClient>();
 
@@ -74,8 +75,6 @@ const TalentLayerProvider = ({ children }: { children: ReactNode }) => {
       );
       currentUser.isAdmin = platform?.address === currentUser?.address;
 
-      setUser(currentUser);
-
       const userHasDelegatedToPlatform =
         process.env.NEXT_PUBLIC_DELEGATE_ADDRESS &&
         userResponse.data.data.users[0].delegates &&
@@ -91,6 +90,9 @@ const TalentLayerProvider = ({ children }: { children: ReactNode }) => {
           userHasDelegatedToPlatform &&
           !userHasReachedDelegationLimit,
       );
+
+      setUser(currentUser);
+
       setLoading(false);
       return true;
     } catch (err: any) {
@@ -116,16 +118,25 @@ const TalentLayerProvider = ({ children }: { children: ReactNode }) => {
   }, [chainId, account.address, talentLayerClient]);
 
   const getWorkerProfile = async (userId: string) => {
-    const response = await getWorkerProfileByOwnerId(userId);
+    try {
+      setLoading(true);
+      const response = await getWorkerProfileByOwnerId(userId);
 
-    if (response.status !== 200) {
-      console.error('Error while fetching worker profile');
-      return;
-    }
+      if (response.status !== 200) {
+        console.error('Error while fetching worker profile');
+        return;
+      }
 
-    const data = await response.json();
-    if (data) {
-      setWorkerProfile(data);
+      const data = await response.json();
+      if (data) {
+        setWorkerProfile(data);
+      }
+    } catch (err: any) {
+      // eslint-disable-next-line no-console
+      console.error(err);
+      return false;
+    } finally {
+      setLoading(false);
     }
   };
