+++ src/modules/Web3mail/components/Web3mailCard.tsx
@@ -1,109 +1,135 @@
-import { ArrowSmallRightIcon } from '@heroicons/react/24/outline';
+import { ArrowSmallRightIcon, ArrowUpRightIcon } from '@heroicons/react/24/outline';
+import Link from 'next/link';
 
 function Web3mailCard() {
   return (
-    <div className='bg-base-200 flex flex-col items-center rounded-2xl p-4 sm:flex-row'>
-      <div className='relative h-[120px] w-[240px] shrink-0 sm:h-[175px] flex items-center'>
-        <svg
-          className='pointer-events-none start-3'
-          xmlns='http://www.w3.org/2000/svg'
-          enableBackground='new 0 0 1920 1080'
-          viewBox='0 0 1920 1080'
-          id='concept-of-data-privacy-and-policy'>
-          <path
-            fill='#2a94f4'
-            d='M1202 1006.8c28.1-13.3 475.7-7 404.5-333-71.2-326-526-287.2-636.8-513.8C858.9-66.5 561.2 30.6 392.9 198.8c-171 171-336.6 945 646.9 596.1 382.7-135.8 143.3 200 143.3 200'
-            opacity='.1'></path>
-          <path
-            fill='#ffbe55'
-            d='M947.9 58c-53.6-8.9-106.7 18.9-129.6 68.2-6.7 14.4-12.5 31.2-16.8 50.7-22.3 101.6-34.6 170.9-127.5 240.3s110.5 284.6 214.4 195.5 279.7 22.9 319.4-43c39.6-65.9-32-137.7 2.5-192.2s67-255.1-27.1-322C1089-11.4 970.4-26.2 947.9 58z'></path>
-          <path
-            fill='#2a94f4'
-            d='M920.2 479.1S695 458.7 739.6 993.5h490.5s46.4-503.3-180.7-514.4H920.2z'></path>
-          <path
-            fill='#fcd2b1'
-            d='M935.3 323.7S918 494 920.4 528.7s107 64.4 126.3 14.9c11.7-30-19.3-219.8-19.3-219.8h-92.1z'></path>
-          <path
-            fill='#ffe3ca'
-            d='M1131.4 112.6s126.3 291.3-64.5 318.1c-49.3 6.9-120.5-47.6-131.6-87.2-9.8-34.7-26.4-8.1-39.1-37.7-17.3-40.4-5.4-156.8 31.7-81.9 27.3 55.1 84.2-102.4 138.7-77.6 54.5 24.8 64.8-33.7 64.8-33.7z'></path>
-          <path
-            fill='#0e538c'
-            d='M843.1 1069.5 664 503.8s-143.5 66-229.7 121.9c0 0 52.7 144.6 91.8 223.8 39.1 79.2 59.3 104.8 124.2 138 65 33.2 192.8 82 192.8 82z'></path>
-          <path
-            fill='#0e538c'
-            d='M843.1 1069.5 664 503.8s155.3-28.6 258-32.5c0 0 40.1 148.6 53.7 235.9 13.6 87.2 11.8 119.8-22.2 184.4-33.9 64.4-110.4 177.9-110.4 177.9z'></path>
-          <path
-            fill='#f9f3ed'
-            d='M459.3 634.1c14.6 39.3 54.5 144.1 85.2 206.3 36.9 74.8 54.4 97.7 115.2 128.9 50.1 25.6 139.8 61.3 175.5 75.2l-165-521.1c-17.7 8.3-132.5 62.2-210.9 110.7z'></path>
-          <path
-            fill='#f9f3ed'
-            d='M935.4 882c31.8-60.5 32.9-89.3 20-171.7-10.6-68.5-38.4-177.2-49-217.8-92.1 5.5-217 27.4-236.2 30.9l165 521.1c21.2-31.9 74-112.7 100.2-162.5z'></path>
-          <path
-            fill='#0e538c'
-            d='M474.3 639.3c17.3 45.9 53 138.7 81.2 195.7 35.4 71.7 50.9 93.2 109.8 123.4 45.7 23.4 124.9 55.3 165.1 71.1L674 535.2c-18 8.4-123.7 58.2-199.7 104.1z'></path>
-          <path
-            fill='#0e538c'
-            d='M924.6 876.3c30.8-58.6 31.1-85 18.8-164.1-9.8-62.8-33.9-159.3-46.2-206.7-88.6 6.2-203.6 26.3-223.2 29.8l156.5 494.2c23.7-36.1 70.2-107.8 94.1-153.2z'></path>
-          <path
-            fill='#d3d3d3'
-            d='m1053.6 1042.1 313.3 34.3c9.7 1.1 18.4-5.9 19.4-15.6l59.6-544.2c1.1-9.7-5.9-18.4-15.6-19.4l-330.8-36.3-61.6 561.7c-1 9.7 6 18.4 15.7 19.5z'></path>
-          <path
-            fill='#f7f7f7'
-            d='m1423.7 496.5-11.7-1.3-312.4-34.2-11.7-1.3c-13.4-1.5-25.4 8.2-26.9 21.6l-58.2 530.8c-1.5 13.4 8.2 25.4 21.6 26.9l299.9 32.9c13.4 1.5 25.4-8.2 26.9-21.6l26.6-242.6 11.7 1.3c13.4 1.5 25.4-8.2 26.9-21.6l28.9-264c1.4-13.5-8.2-25.5-21.6-26.9z'></path>
-          <path
-            fill='#cce9ff'
-            d='M886.8 794.9V706c0-28.4 26.2-49.6 47.4-49.6h51c35.7 0 54.7 21.1 54.7 49.6v88.9h43.7V706c0-52.5-38.6-93.3-98.4-93.3h-51c-45.2 0-91.1 40.8-91.1 93.3v88.9h43.7z'></path>
-          <path
-            fill='#ffbe55'
-            d='M1105.4 1022.3c0 24.1-14.6 42.3-38.6 42.3H867.1c-24.1 0-45.9-18.2-45.9-42.3V808c0-24.1 21.9-27.7 45.9-27.7h199.7c24.1 0 38.6 3.6 38.6 27.7v214.3z'></path>
-          <path
-            fill='#ffbe55'
-            d='M1002.6 891.9c0-20.4-16-37.2-36.4-37.2s-35.7 16.8-35.7 37.2c0 14.6 7.3 26.2 21.9 32.8v68.5c0 9.5 5.1 16.8 14.6 16.8s14.6-7.3 14.6-16.8v-68.5c14.5-6.6 21-18.3 21-32.8z'></path>
-          <path
-            d='M1002.6 891.9c0-20.4-16-37.2-36.4-37.2s-35.7 16.8-35.7 37.2c0 14.6 7.3 26.2 21.9 32.8v68.5c0 9.5 5.1 16.8 14.6 16.8s14.6-7.3 14.6-16.8v-68.5c14.5-6.6 21-18.3 21-32.8z'
-            opacity='.13'></path>
-          <path
-            fill='#ffe3ca'
-            d='M773.4 987.2 832 972s7.1-27.9 20.5-38.5 29.4-18.6 36.6-17.7c7.1.9-8 20.5-15.2 29.4s3.6 11.6 32.1-8 52.6-25.9 59.7 14.2c7.1 40.2-33.9 58.9-83.8 58.9s-89.1 28.1-89.1 28.1-74.4-6.6-19.4-51.2z'></path>
-          <path
-            fill='#2a94f4'
-            d='M526.2 849.5s-21.9 1.2-41.7 40.8c-19.8 39.6 5.5 229.1 350.8 154.2l-22.7-71.8c-.1 0-266.7-80.6-286.4-123.2z'></path>
-          <path
-            fill='#ffe3ca'
-            d='m1162.9 987.2-58.6-15.2s-7.1-27.9-20.5-38.5-29.4-18.6-36.6-17.7c-7.1.9 8 20.5 15.2 29.4s-3.6 11.6-32.1-8-52.6-25.9-59.7 14.2c-7.1 40.2 33.9 58.9 83.8 58.9s89.1 28.1 89.1 28.1 74.4-6.6 19.4-51.2z'></path>
-          <path
-            fill='#2a94f4'
-            d='M1409.6 849.5s21.9 1.2 41.7 40.8c19.8 39.6-5.5 229.1-350.8 154.2l22.7-71.8c.1 0 266.6-80.6 286.4-123.2z'></path>
-        </svg>
-      </div>
-      <div className='mt-6 grow sm:mt-0'>
-        <div className='pb-4 text-center sm:pb-0 sm:text-left'>
-          <p className='font-heading text-xl font-semibold leading-normal mb-2 text-base-content opacity-90'>
-            <span>
-              {' '}
-              configure private email alerts <span className='text-3xl'>🪄</span>
-            </span>
-          </p>
-          <p className='font-alt text-sm font-normal leading-normal max-w-sm text-base-content opacity-70'>
-            <span>
-              {' '}
-              your email will never be stored by us, BuilderPlace, or TalentLayer. private email is
-              powered by iExec Web3Mail - allowing us to send you email alerts without your data
-              being disclosed to us.
-            </span>
-          </p>
-          <div className='mt-3'>
-            <a
-              href='https://medium.com/iex-ec/introducing-privacy-enhancing-marketing-the-future-of-marketing-in-web3-429283e7c7f0'
-              target='_blank'
-              className='px-3 py-1 border border-info rounded-md hover:bg-base-200 text-base-content hover:text-base-content inline-flex items-center text-xs'>
-              <span>learn more</span>
-              <ArrowSmallRightIcon width='16' height='16' className='ml-2' />
-            </a>
+    <>
+      <div className='bg-base-200 flex flex-col items-center rounded-2xl p-4 sm:flex-row'>
+        <div className='relative h-[120px] w-[240px] shrink-0 sm:h-[175px] flex items-center'>
+          <svg
+            className='pointer-events-none start-3'
+            xmlns='http://www.w3.org/2000/svg'
+            enableBackground='new 0 0 1920 1080'
+            viewBox='0 0 1920 1080'
+            id='concept-of-data-privacy-and-policy'>
+            <path
+              fill='#2a94f4'
+              d='M1202 1006.8c28.1-13.3 475.7-7 404.5-333-71.2-326-526-287.2-636.8-513.8C858.9-66.5 561.2 30.6 392.9 198.8c-171 171-336.6 945 646.9 596.1 382.7-135.8 143.3 200 143.3 200'
+              opacity='.1'></path>
+            <path
+              fill='#ffbe55'
+              d='M947.9 58c-53.6-8.9-106.7 18.9-129.6 68.2-6.7 14.4-12.5 31.2-16.8 50.7-22.3 101.6-34.6 170.9-127.5 240.3s110.5 284.6 214.4 195.5 279.7 22.9 319.4-43c39.6-65.9-32-137.7 2.5-192.2s67-255.1-27.1-322C1089-11.4 970.4-26.2 947.9 58z'></path>
+            <path
+              fill='#2a94f4'
+              d='M920.2 479.1S695 458.7 739.6 993.5h490.5s46.4-503.3-180.7-514.4H920.2z'></path>
+            <path
+              fill='#fcd2b1'
+              d='M935.3 323.7S918 494 920.4 528.7s107 64.4 126.3 14.9c11.7-30-19.3-219.8-19.3-219.8h-92.1z'></path>
+            <path
+              fill='#ffe3ca'
+              d='M1131.4 112.6s126.3 291.3-64.5 318.1c-49.3 6.9-120.5-47.6-131.6-87.2-9.8-34.7-26.4-8.1-39.1-37.7-17.3-40.4-5.4-156.8 31.7-81.9 27.3 55.1 84.2-102.4 138.7-77.6 54.5 24.8 64.8-33.7 64.8-33.7z'></path>
+            <path
+              fill='#0e538c'
+              d='M843.1 1069.5 664 503.8s-143.5 66-229.7 121.9c0 0 52.7 144.6 91.8 223.8 39.1 79.2 59.3 104.8 124.2 138 65 33.2 192.8 82 192.8 82z'></path>
+            <path
+              fill='#0e538c'
+              d='M843.1 1069.5 664 503.8s155.3-28.6 258-32.5c0 0 40.1 148.6 53.7 235.9 13.6 87.2 11.8 119.8-22.2 184.4-33.9 64.4-110.4 177.9-110.4 177.9z'></path>
+            <path
+              fill='#f9f3ed'
+              d='M459.3 634.1c14.6 39.3 54.5 144.1 85.2 206.3 36.9 74.8 54.4 97.7 115.2 128.9 50.1 25.6 139.8 61.3 175.5 75.2l-165-521.1c-17.7 8.3-132.5 62.2-210.9 110.7z'></path>
+            <path
+              fill='#f9f3ed'
+              d='M935.4 882c31.8-60.5 32.9-89.3 20-171.7-10.6-68.5-38.4-177.2-49-217.8-92.1 5.5-217 27.4-236.2 30.9l165 521.1c21.2-31.9 74-112.7 100.2-162.5z'></path>
+            <path
+              fill='#0e538c'
+              d='M474.3 639.3c17.3 45.9 53 138.7 81.2 195.7 35.4 71.7 50.9 93.2 109.8 123.4 45.7 23.4 124.9 55.3 165.1 71.1L674 535.2c-18 8.4-123.7 58.2-199.7 104.1z'></path>
+            <path
+              fill='#0e538c'
+              d='M924.6 876.3c30.8-58.6 31.1-85 18.8-164.1-9.8-62.8-33.9-159.3-46.2-206.7-88.6 6.2-203.6 26.3-223.2 29.8l156.5 494.2c23.7-36.1 70.2-107.8 94.1-153.2z'></path>
+            <path
+              fill='#d3d3d3'
+              d='m1053.6 1042.1 313.3 34.3c9.7 1.1 18.4-5.9 19.4-15.6l59.6-544.2c1.1-9.7-5.9-18.4-15.6-19.4l-330.8-36.3-61.6 561.7c-1 9.7 6 18.4 15.7 19.5z'></path>
+            <path
+              fill='#f7f7f7'
+              d='m1423.7 496.5-11.7-1.3-312.4-34.2-11.7-1.3c-13.4-1.5-25.4 8.2-26.9 21.6l-58.2 530.8c-1.5 13.4 8.2 25.4 21.6 26.9l299.9 32.9c13.4 1.5 25.4-8.2 26.9-21.6l26.6-242.6 11.7 1.3c13.4 1.5 25.4-8.2 26.9-21.6l28.9-264c1.4-13.5-8.2-25.5-21.6-26.9z'></path>
+            <path
+              fill='#cce9ff'
+              d='M886.8 794.9V706c0-28.4 26.2-49.6 47.4-49.6h51c35.7 0 54.7 21.1 54.7 49.6v88.9h43.7V706c0-52.5-38.6-93.3-98.4-93.3h-51c-45.2 0-91.1 40.8-91.1 93.3v88.9h43.7z'></path>
+            <path
+              fill='#ffbe55'
+              d='M1105.4 1022.3c0 24.1-14.6 42.3-38.6 42.3H867.1c-24.1 0-45.9-18.2-45.9-42.3V808c0-24.1 21.9-27.7 45.9-27.7h199.7c24.1 0 38.6 3.6 38.6 27.7v214.3z'></path>
+            <path
+              fill='#ffbe55'
+              d='M1002.6 891.9c0-20.4-16-37.2-36.4-37.2s-35.7 16.8-35.7 37.2c0 14.6 7.3 26.2 21.9 32.8v68.5c0 9.5 5.1 16.8 14.6 16.8s14.6-7.3 14.6-16.8v-68.5c14.5-6.6 21-18.3 21-32.8z'></path>
+            <path
+              d='M1002.6 891.9c0-20.4-16-37.2-36.4-37.2s-35.7 16.8-35.7 37.2c0 14.6 7.3 26.2 21.9 32.8v68.5c0 9.5 5.1 16.8 14.6 16.8s14.6-7.3 14.6-16.8v-68.5c14.5-6.6 21-18.3 21-32.8z'
+              opacity='.13'></path>
+            <path
+              fill='#ffe3ca'
+              d='M773.4 987.2 832 972s7.1-27.9 20.5-38.5 29.4-18.6 36.6-17.7c7.1.9-8 20.5-15.2 29.4s3.6 11.6 32.1-8 52.6-25.9 59.7 14.2c7.1 40.2-33.9 58.9-83.8 58.9s-89.1 28.1-89.1 28.1-74.4-6.6-19.4-51.2z'></path>
+            <path
+              fill='#2a94f4'
+              d='M526.2 849.5s-21.9 1.2-41.7 40.8c-19.8 39.6 5.5 229.1 350.8 154.2l-22.7-71.8c-.1 0-266.7-80.6-286.4-123.2z'></path>
+            <path
+              fill='#ffe3ca'
+              d='m1162.9 987.2-58.6-15.2s-7.1-27.9-20.5-38.5-29.4-18.6-36.6-17.7c-7.1.9 8 20.5 15.2 29.4s-3.6 11.6-32.1-8-52.6-25.9-59.7 14.2c-7.1 40.2 33.9 58.9 83.8 58.9s89.1 28.1 89.1 28.1 74.4-6.6 19.4-51.2z'></path>
+            <path
+              fill='#2a94f4'
+              d='M1409.6 849.5s21.9 1.2 41.7 40.8c19.8 39.6-5.5 229.1-350.8 154.2l22.7-71.8c.1 0 266.6-80.6 286.4-123.2z'></path>
+          </svg>
+        </div>
+        <div className='mt-6 grow sm:mt-0'>
+          <div className='pb-4 text-center sm:pb-0 sm:text-left'>
+            <p className='font-heading text-xl font-semibold leading-normal mb-2 text-base-content opacity-90'>
+              <span>
+                {' '}
+                configure private email alerts <span className='text-3xl'>🪄</span>
+              </span>
+            </p>
+            <p className='font-alt text-sm font-normal leading-normal max-w-sm text-base-content opacity-70'>
+              <span>
+                {' '}
+                your email will never be stored by us, BuilderPlace, or TalentLayer. private email
+                is powered by iExec Web3Mail - allowing us to send you email alerts without your
+                data being disclosed to us.
+              </span>
+            </p>
+            <div className='mt-3'>
+              <a
+                href='https://medium.com/iex-ec/introducing-privacy-enhancing-marketing-the-future-of-marketing-in-web3-429283e7c7f0'
+                target='_blank'
+                className='px-3 py-1 border border-info rounded-md hover:bg-base-200 text-base-content hover:text-base-content inline-flex items-center text-xs'>
+                <span>learn more</span>
+                <ArrowSmallRightIcon width='16' height='16' className='ml-2' />
+              </a>
+            </div>
           </div>
         </div>
       </div>
-    </div>
+      <div className='border-b border-info pb-6'>
+        <label className='block'>
+          <div className='mb-2 ml-0.5'>
+            <div className='flex justify-between'>
+              <div>
+                <p className='font-heading text-base-content font-medium leading-none'>
+                  1. setup your email
+                </p>
+                <p className='font-sans text-xs font-normal leading-normal text-base-content mt-0.5'>
+                  configure your preferences to be notified every time an important event happens on
+                  your profile.
+                </p>
+              </div>
+            </div>
+          </div>
+        </label>
+        <Link
+          href='/web3mail'
+          className='grow px-5 py-2 rounded-xl bg-info text-base-content hover:bg-base-200 inline-flex align-center justify-center '>
+          <span>manage your email</span>
+          <ArrowUpRightIcon width='16' height='16' className='ml-2 mt-1' />
+        </Link>
+      </div>
+    </>
   );
 }
 export default Web3mailCard;
